+++ script/module/config.js
@@ -1,0 +1,8 @@
+const config = {
+    // 渲染的自定义样式
+    styles: [
+        'font-size',
+    ],
+};
+
+export { config };